refactor: Extract helper
